Working on the 'logo'
There are some fields in structure that are injected into
_structure.html.haml by methods. To serialize these as JSON in the view,
(for bootstrapping), I thought to use,

  self.to_json(:methods => [:some_method])

However, this means making methods for a bunch of things. For example,
in structure.rb, there is now a,

  def logo_present
    return this.logo.present?
  end

Which is probably awful, because I will need like a ton of these
methods.

So I'm taking a break now to discus this with Nima, and see if there is
a better way to serialize the models for use in bootstrapping.