Included the GoogleMapsView
Creates a view and templates based on a jsfiddle:

  http://jsfiddle.net/rubyjedi/WTkc9/1/

And them moved things around in the templates and rails views until
everything worked. Now the map container shows, but no map. I've removed
the original google map, just in case it is interfering.

Installed the GMV in the Widgets layout.