Converting backbone.googlemaps
marionette views are built on top of backbone views, and so are backbone
plugin views. Marionette generally is adding the sort of functionality
that the plugin writers added, but is consistent across views.

Rather than having a bunch of plugins that all handle 'being a view'
differently, I find it productive to try to convert.

So I'm just in the process of making a google_maps_view that extends
from marionette's composite view.