First Contact!
The paginator shouted out to the layout, which in turn informed the map
widget that something had happened.

This needs to be cleaned up a little: the paginator should package all
the information that the map widget will need in an event object, and
the layout should simply pass on the event object. The knowledge of
which method to call should come from within the map widget.

So, paginator says "I updated the results!" and map widget responds in
its own way. Hmm... I think this means we need the map widget to
register as a listener for a particular kind of event: it should say
"Hey layout, if you hear any 'paginator:pageChanged', let me know and
give me the info."