Sorting works now
The problem was the collection was not wiping itself before getting the
results (which is smart pagniating). So some of the sorted results
coming in were already in the collection, and thus would not appear in
the correct order.

Now, when changing sorting methods we 'reset' the collection (but not
the cache, you see).