Bug fix!
The latlngbounds are stored as CVS in the URI, and thus also as CSV in the
paginated_collection.server_api, but we usually don't want CSV everywhere
else. So, when paginated_collection_view is announcing the latlngbounds of
the current result-set to the map, so that it can show the right view port,
it needs to receive the latlngbounds as a nice object.