Structure Item -- Showable
Finished some checklist items:

StructureView will become a layout, and it will house composite views
constructed from its data-members.

Add the ability to have multiple Collections in a single ESView (a layout
and regions, with clickable hooks on the structure view).

The clickable area for expanding the view should be small.

For now the structureView regions will have data from the structure:
subjects and places.

Card 213