Structure Item -- Showable
Pulled a "relational accordion item view" out of the structure view. This
functionality had little to do with the structure.

Also, changed things so that we can used 'data-attributes' on the button
for the relational accordion view to store a list of model attributes to
use as the "composite" portion of the composite view... yes!