Marker Bounce
Changed the language used in the relational_accordion_item_view, and
accordion_item_view, to jive with that used in the event_layout. That
is, we are using data-behavior and data-model, with the name of the
relation's model (ie for courses, we have course). The layouts themselves
are attached to the event_layout at the relevant data-type node, with
the type matching the name of the view.