Directory Structure
I rearranged the files, so that view and templates directories
both have many layers of nested sub directories. This necessitated
the addition of namespace beyond just App.Views and App.Models

Now, if a view resides in views/filtered_search/map it will be a
member of the module Views.FilteredSearch.Map, which provides the
method `templateDirname` pointing to the file directory. The
template corresponding to this view will be in the same path,
but in the templates directory instead of the views directory.

For the relational_accordion_item_view, we had to create a method
that, when executed in the context of the implementing class, gets
the module that the relation would exist in. For a relation Widgets
on Views.MyModule.MyObject, we would expect Views.MyModule.MyObject.Widgets
to be the correct module.

Any class extending from relational_accordion_item_view must bind
its Module to the getModuleForRelation method in initialize, like so:

  this.getModuleForRelation = _.bind(this.getModuleForRelation, Module);