CRM -- paginated table view
BOOM! We have paginators on the top and bottom, we have results,
we have everything.

At the moment the collection starts at page 2, because we didn't have
bootstrap data and I couldn't figure out how to make it query for the
first page automatically, so that is just to remind me to fix that
later.