Refactor -- onMarkerFocus
Two problems. onMarkerFocus was defined in the super class, which is wrong
because that should be a life cycle method. So I bound a markerFocus method
to the 'focus' event directly in the constructor. Now child classes can
override onMarkerFocus with their own behaviour (such as showing the
infobox).

The other problem, which was actually a bug, was that one of the google
map views was not registering its markerViews to do child bubbling.