Overwrote fetch in user_profiles_collection
Fetch grabs the entire default collection from the server, and then
updates all the models to reflect the changes.

Our collection checks whether there are more changes to come,
and only updates the model when all the changes are through. Later
the collection won't even fetch the changes, it will only fetch
a boolean value that will allow it to decide whether or not
to fetch.