Currently editing field is immune to update
The value in the currently editing field won't be updated when a mass
update goes through.

Ultimately the value may still be updated, though:

for example, if I add "bob" as a tag to all students, and while that
is processing I add "jill" as a tag to one student, then the one
student will have both tags. The same is not true for text fields
like "first_name". This is not implemented yet.