There is a growing evil in this place
Collection View needs to keep track of "unfinished" editables. In generaly, it
needs to know which editable is "currently editing" so that it can do some
stuff that could not be the responsibility of the editable itself, or the editing
row. Meta stuff.

The problem is, this means it has a reference to the currently_editing row...
which is fine! The row is a child of the collection, so references are allowed.
Except that the collection is, for convenience, calling methods
like "finishEditing" that directly affect the editables. The Collection shouldn't
know about the editables, so this is tight coupling.

I propose to make some methods of the row _private_. Those methods that deal
with the editables deep inside the row will be private. To activate such
functionality, the collection should have to announce its own state... and just
hope for a sane response.

This has been irking me.