Used the Null Object Pattern to prevent crazy
When you rebrand one app as another, the original app is effectively destroyed.
Since ages past, we had had this lurking issue that our site always loads all
modules from all apps. This allows us to orchestrate inheritance, and we control
the load order with manifests.

However, the result this time was that the FilteredSearch app was _always_ being
re-branded, even when we were legitimately on the FilteredSearch page. Right away
I tried to fix this by replacing apps with NullApplications if they're app root
it not on the page. This was a problem right away, because we need CoursAvenue
app to always load.

For now I'm using the null object pattern, but limiting it to apps that are built
via rebranding. Maybe later I will fix things so that only the current app and
the coursavenue app are available (would have to make sure no sharing is going
on, though, so that would probably mean one of the other apps would have to
change too...).