Force the Courses tab to show
Since adding the relational views to the tabs, and showing tabs, are both
triggered by a click, and since adding the tab involves an async fetch, I've
had to use promise magic to force the Courses tab to be "active" by default.

Probably there is a better way, and the best way would be to move the tabbing
logic into this class (where it belongs).