Added Teachers relation
Added teachers to the serializer and backbone relational model.

Changed relational_tab_layout so that it either fetches the models if they were
side-loaded, or straight-up uses them if they were bootstrapped. This is
because I'm not side-loading the teachers (which, theoretically, I should be)
but it is also a good thing to have, so I'm glad. ^o^//