Split tab_manager module
There is a folder, tab_manager, and a subfolder flavors. The same structure is
mirrored in the templates dir, and there is no need for manifest.

In the tab_manager folder there are two files, tab_manager.js and
tab_manager_view.js which define the module and its generic business, and the
view that the module serves. The flavors folder contains submodules that
define extensions of the base view.

The hope is that the flavors will be really minimal and not full of gotchas.
What I would like most would be if the flavors only override template, and
use initialize to add behaviors (I'm thinking, can we make behaviors polymorphic
so that they can accept views as well as DOM elements).