Sort of started to try to do the filtering
I got stuck on a conceptual problem. I kept thinking it should be possible
to do something like,

`Structure.where(id: 39).with("planning").where(audience_ids: "!")`

the `with` in there is like saying "give me back a relation that has
been secretly hacked up so that it matches the following condition".
You know? I want the relation to contain what we are looking for,
rather than having to do it in two steps and return some json or
something. Turns out this might be more complicated than I thought,
and it is definitely not built in. It probably wouldn't be efficient,
either, so I have no idea why I am so focused on it.