Switched to using a structure
OK so during the daedalus stuff, the manner in which the server communicated the structure
to the client became more and more fractured and weird. By the end we were passing
all the relations separately, because of the pain involved in doing things right. Now,
we are doing things right. We only pass out the structure and its places (and I think
we can get rid of the places too).