Changes to filter_breadcrumbs_view
The filter_breadcrumbs_view now accepts an object mapping between the
internal names of the filters it cares about, and the visible names.

Debounced the render function as it was sometimes being called
10 times in a row with the same args.

Added a method that calls addBreadcrumb multiple times, so that
the Structure can announce all its filters at once.