Event Layout Edge Case
Previously we changed event layout so that, if the event layout was
its own master region it could declare its events and DOM events all
in the same hash. This is used in the CRM page.

This time the layout's master region is housing a view class that
descends from EventLayout, and the above changes caused a conflict.
This commit resolves that conflict.