Implement processing of charge disputes.
We process three charge dispute related events: at the creation of the
dispute, when the funds are withdrawn and when the funds are reinstated.

When the dispute is created, we pause the subscription until we have
further information. When the funds are withdrawn, we cancel the
subscription. Finally, when the funds are reinstated we resume the
subscription.