Only show structure website if it's actually set.
Currently, the `website_link` uses the `link_to` helper method, and the
method doesn't fail when the passed link is `nil`. Instead, it just
returns a link without anything in the `href` attribute.

Apparently Chrome (and maybe other browser) decide to fill the `href`
with the current URL.

We prevent this by not showing the website link if the website is not
set.