Prefix delegation of name method in vouchers.
In the voucher model, we have an `attr_accessor` for name, used in the
creation form. We also delegated the `name` method to the parent gift
certificate which caused ambiguity when the `voucher.name` was used. We
fixed that by prefixing the delegation: we now use
`voucher.gift_certificate_name` when we want to access the name.