Raise a RoutingError when the route is not matched.
When trying to access a random route, Rails would just render an
`about:blank` page instead of the standard error page or the redirection
to the `root_path`. To fix this, we match every path that is not
previously matched in the route file and point to an endpoint that will
correctly raise the right error.