Add comments_count to Comment.
This field is also needed in the Comment table because the comments
association is a polymorphic one, and the reverse association table
include comments. This means that every time a `Comment::Review` is
created, we try to increment the associated model (here
`Comment::Reply`) `comments_count` field. In this case, the
`comments_count` in `Comments` will / should remain unused.

[1]: http://yerb.net/blog/2014/03/13/three-easy-steps-to-using-counter-caches-in-rails/